+++ app/src/test/java/br/com/concrete/tentacle/base/BaseTest.kt
@@ -4,6 +4,7 @@ import android.app.Instrumentation
 import br.com.concrete.tentacle.data.network.ApiService
 import br.com.concrete.tentacle.di.networkModule
 import br.com.concrete.tentacle.di.repositoryModule
+import br.com.concrete.tentacle.di.sharedPreferencesModule
 import br.com.concrete.tentacle.di.viewModelModule
 import br.com.concrete.tentacle.rules.RxImmediateSchedulerRule
 import io.reactivex.Flowable
